ghc: Upgrade to GHC 7.8.4.
A few major thing changed with 7.10, making upgrading beyond 7.8 much
more harder than it looks:
https://ghc.haskell.org/trac/ghc/wiki/Migration/7.10

This is already quite a leap forward. Many thing changed, starting with
cabal 1.18 which is likely to be a hassle for haskell recipes.

Unfortunately, GHC 6 does not play well with libc 2.25 (maybe
before...) and is afflicted segfaults in its RTS, and does not respect
the ABI for stack alignment...

TL;DR: Maintaining a GHC 6 within a modern runtime environment is fun
and all, but life is too short...

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>