haskell: ghc local package database
The database should be generated in ${B} not ${S}.
This also explains why we needed to add the depends on do_unpack when
only the recipe-sysroot is used to build the initial database.

The ticket shows that do_configure is failing to find the database,
probably because ${S} is not ready yet as it is created during
do_unpack.
From base.conf:
do_unpack[dirs] = "${WORKDIR}"
roughly:do_unpack[cleandirs] = "${S}"
This guaranties that ${S} is created and empty.

do_update_local_pkg_database[dirs] = "${B}"
Will guaranty that ${B} is created and should remove the need to depend
on do_unpack to create ${S} for the database to be written into.
bitbake.conf defines B = ${S} by default though, so that's still not
quite fitting.

do_{fetch,unpack,patch} are not defined with externalsrc.bbclass, which
breaks Haskell recipes in that context.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1317