haskell: ghc local package database
The database is in ${B}, not ${S}. This might explain why we needed to
add the depends on do_unpack when only the recipe-sysroot is used to
build the initial database.
The ticket also shows that do_configure is failing to find the database,
probably because ${S} != ${B}.

do_{fetch,unpack,patch} are not defined with externalsrc.bbclass, which
breaks Haskell recipes in that context.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1317