haskell: use sysroot package database
The sysroot already has a package database that gets populated with the
sysroot. With a minor quirks for target classes, it should rather use
this since task dependency will let us maintain the database sanity.
That way should S be hit by an earthquake (e.g, git clean...) bitbake
can rely safely on the sstate and the database in the sysroot is still
valid for do_compile.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>